requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.9
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dataclasses==0.6
deepsegment==2.3.1
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.1.0
flatbuffers==25.2.10
gast==0.6.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud==0.34.0
google-cloud-speech==2.32.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h5py==3.13.0
idna==3.10
importlib-metadata==6.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
murmurhash==1.0.12
mypy_extensions==1.1.0
namex==0.0.9
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pathlib==1.0.1
pipdeptree==2.26.1
preshed==3.0.9
progressbar2==4.5.0
proto-plus==1.26.1
protobuf==6.31.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydload==1.0.9
Pygments==2.19.1
pyrsistent==0.19.3
python-dotenv==1.1.0
python-utils==3.9.1
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
seqeval==0.0.3
seqtag-keras==1.0.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.6
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.66.1
typer==0.15.4
typing-inspect==0.8.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wasabi==1.1.1
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.15.0
